+++ core/documentation/source/basic/changelog.rst
@@ -1,6 +1,6 @@
-######################
+#########
 ChangeLog
-######################
+#########
 
 **v 1.2.2** (Planned)
 
@@ -11,6 +11,8 @@ ChangeLog
 - Now modules can registers theirself to add their package path to extend core classes, helpers, etc.. inside their "extend" folder, such as **application/modules/mymodule/extend/helper/some_helper.php**
 - Modules now can also implement a dedicated menu on the administration if they registers as "package" on the settings. Please read more on the documentation
 - Fixed a bug on the settings "delete" method
+- New categories() method on the record object to extract an array of the record categories
+- Speed improvements on the related_records() function of the frontend helper
 
 
 **v 1.2.1** (2012-05-02)